+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+v1.0.4
+ - Fix versioning problem (@lucas.oliveira)
+ - Changed `gitlab-backup.sh` to keep all backups (@lucas.oliveira)
+
 v1.0.3
  - Fix only numbers tag bug. (@lucas.oliveira)
  - Add version to log. (@lucas.oliveira)